src/main/java/com/exchange/service/CurrencyServiceImpl.java
package com.exchange.service;

import com.exchange.domain.Currency;
import com.exchange.repository.CurrencyRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class CurrencyServiceImpl implements CurrencyService {
    private CurrencyRepository currencyRepository;

    @Autowired
    public CurrencyServiceImpl(final CurrencyRepository currencyRepository) {
        this.currencyRepository = currencyRepository;
    }

    @Override
    public Iterable<Currency> getAllCurrencies() {
        return currencyRepository.findAll();
    }
}
